requirements.txt
accelerate==0.30.1
adjustText==1.1.1
appdirs==1.4.4
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
Ciw==3.1.0
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
filelock==3.13.1
fonttools==4.47.2
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.41
gmpy2==2.1.2
gpytorch==1.11
graphviz==0.20.3
gymnasium==0.29.1
higher==0.2.1
huggingface-hub==0.23.0
idna==3.6
jaxtyping==0.2.28
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
linear-operator==0.5.2
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
optree==0.11.0
packaging==23.2
pandas==2.1.4
Pillow==9.4.0
pip==23.3.2
protobuf==4.25.2
psutil==5.9.7
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.1.post1
scipy==1.12.0
sentry-sdk==1.39.2
setproctitle==1.3.3
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
sympy==1.12
threadpoolctl==3.3.0
torch==2.1.2
torchaudio==2.1.2
torchopt==0.7.3
torchvision==0.16.2
tqdm==4.66.1
triton==2.1.0
typeguard==2.13.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wandb==0.16.2
wheel==0.42.0
